backend/requirements.txt
uwsgi==2.0
django==1.8.5
djangorestframework==3.3.0
requests==2.9.1
dj_database_url==0.4.0
psycopg2==2.6.1
envtpl==0.4.1
django-cors-headers
django-filter
gunicorn
python-slugify
djangorestframework-jwt
rest_condition